cleanup and added BBWS and millerRabin to RSA. Also works with 64 bit ints now as well

+++ rsa_ykaravas/include/PrimeClass.h
@@ -1,24 +1,25 @@
 #ifndef PRIMECLASS_H
 #define PRIMECLASS_H
 
+#include <cstdint>
 
 class PrimeClass {
 
 public:
 
-    PrimeClass(long long int num );
+    PrimeClass(uint64_t num );
 
-    inline void findFactor(long long int start);
+    inline void findFactor(uint64_t start);
     void bruteForce();
 
-    long long int get_n() const {return n;}
-    long long int get_p() const {return p;}
-    long long int get_q() const {return q;}
-    long long int get_sqrt_n() const {return n_sqrt;}
+    uint64_t get_n() const {return n;}
+    uint64_t get_p() const {return p;}
+    uint64_t get_q() const {return q;}
+    uint64_t get_sqrt_n() const {return n_sqrt;}
 
-    long long int n;
-    long long int p;
-    long long int q;
-    long long int n_sqrt;
+    uint64_t n;
+    uint64_t p;
+    uint64_t q;
+    uint64_t n_sqrt;
 };
 #endif
